Move dbs to to its own file
