+++ q2_moshpit/eggnog/_method.py
@@ -17,7 +17,6 @@ from q2_types_genomics.genome_data import SeedOrthologDirFmt, OrthologFileFmt
 from q2_types_genomics.reference_db import EggnogRefDirFmt
 from q2_types.feature_data import DNAFASTAFormat
 from q2_types_genomics.reference_db import DiamondDatabaseDirFmt
-from .._utils import run_command
 from q2_types_genomics.feature_data import (
     OrthologAnnotationDirFmt, MAGSequencesDirFmt
 )
@@ -132,28 +131,3 @@ def _annotate_seed_orthologs_runner(seed_ortholog, eggnog_db, sample_label,
         cmds.append('--dbmem')
 
     subprocess.run(cmds, check=True)
-
-
-def fetch_eggnog_db() -> EggnogRefDirFmt:
-    """
-    Downloads eggnog reference database using the
-    `download_eggnog_data.py` script from eggNOG. Here, this
-    script downloads 3 files amounting to 47Gb in total.
-    """
-
-    # Initialize output objects
-    eggnog_db = EggnogRefDirFmt()
-
-    # Define command.
-    # Meaning of flags:
-    # y: Answer yest to all prompts thrown by download_eggnog_data.py
-    # D: Do not download the Diamond database
-    # data_dir: location where to save downloads
-    cmd = [
-        "download_eggnog_data.py", "-y", "-D",
-        "--data_dir", str(eggnog_db.path)
-    ]
-    run_command(cmd)
-
-    # Return objects
-    return eggnog_db
